# RFC: E notation BigDecimal parser