Related to : #9580 

This implements a first pass at this (not stack based parser, to avoid needing a stack, just a translation of the parse table into logic)

PR which is still a work in progress and would benefit from your inputs should we wish to progress with this

created with the productions below (syntax below works in http://hackingoff.com/compilers/ll-1-parser-generator)

```
Scientific -> OptionalSign Number
OptionalSign -> s |
Number -> Digits OptionalFractional OptionalExponent | Fractional OptionalExponent
OptionalFractional -> '.' OptionalDigits  
Fractional -> '.' Digits
OptionalExponent -> e OptionalSign Digits | 
OptionalDigits -> Digits |  
Digits -> Digit RepeatDigit
RepeatDigit -> Digit |  
Digit -> n 
```