@stevegeek 2020-07-07
@sija things are slower at the moment:

```
require "benchmark"

Benchmark.ips do |x|
  x.report("ex1 - long") do
    BigDecimal.new("+0829756293456064502345702345.283457863284759365e-7364756823")
  end

  x.report("ex2 - long error") do
    BigDecimal.new("138479783434.927347264629873647269348762e82347978e")
  rescue
    #
  end

  x.report("ex3 - short") do
    BigDecimal.new("1.1")
  end

  x.report("ex3 - short error") do
    BigDecimal.new("1.1-")
  rescue
    #
  end
end
```

On current master:

```
MRc:crystal stephen$ ./bin/crystal run --release bdp.cr 

       ex1 - long 781.23k (  1.28µs) (± 0.90%)  672B/op   4.19× slower
 ex2 - long error 434.36k (  2.30µs) (± 1.56%)  785B/op   7.54× slower
      ex3 - short   3.27M (305.53ns) (± 1.15%)  464B/op        fastest
ex3 - short error 578.83k (  1.73µs) (± 1.59%)  464B/op   5.65× slower
```

this: 

```
MRc:crystal stephen$ bin/crystal run --release bdp.cr
       ex1 - long 398.11k (  2.51µs) (± 1.12%)   1.2kB/op   4.99× slower
 ex2 - long error 275.56k (  3.63µs) (± 1.20%)  1.27kB/op   7.20× slower
      ex3 - short   1.99M (503.74ns) (± 1.52%)    784B/op        fastest
ex3 - short error 405.12k (  2.47µs) (± 1.45%)  1.04kB/op   4.90× slower
```