@lbguilherme 2022-03-11
Are there cases that this new parser can handle and the original parser can't?